Desktop/커서ai 모음/massage_shop_web/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1 